+++ includes/parserhooks/Maps_Geodistance.php
@@ -32,22 +32,9 @@ class MapsGeodistance extends ParserHook {
 	 * @return array
 	 */
 	protected function getParameterInfo( $type ) {
-		global $egMapsDistanceUnit, $egMapsDistanceDecimals, $egMapsAvailableGeoServices, $egMapsDefaultGeoService; 
+		global $egMapsDistanceUnit, $egMapsDistanceDecimals;
 		
 		$params = [];
-
-		$params['mappingservice'] = [
-			'default' => '',
-			'values' => MapsMappingServices::getAllServiceValues(),
-			'tolower' => true,
-		];
-
-		$params['geoservice'] = [
-			'default' => $egMapsDefaultGeoService,
-			'aliases' => 'service',
-			'values' => $egMapsAvailableGeoServices,
-			'tolower' => true,
-		];
 
 		$params['unit'] = [
 			'default' => $egMapsDistanceUnit,
@@ -62,13 +49,11 @@ class MapsGeodistance extends ParserHook {
 		$params['location1'] = [
 			'type' => 'mapslocation', // FIXME: geoservice is not used
 			'aliases' => 'from',
-			'dependencies' => [ 'mappingservice', 'geoservice' ],
 		];
 
 		$params['location2'] = [
 			'type' => 'mapslocation', // FIXME: geoservice is not used
 			'aliases' => 'to',
-			'dependencies' => [ 'mappingservice', 'geoservice' ],
 		];
 
 		// Give grep a chance to find the usages: